LayoutTests/css-dark-mode/color-scheme-css-parse-expected.txt


PASS Test valid values
FAIL Test valid values with <custom-ident> assert_equals: expected "only" but got "normal"
PASS Test invalid values
